Move project context into main
